# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "BrasaFut API - Documentação Oficial",
    "description": "API profissional de futebol de alta performance, com suporte a campeonatos brasileiros e internacionais, dados em tempo real, estatísticas de partidas e webhooks.",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "name": "x-api-key",
        "in": "header",
        "description": "Chave de acesso obtida via /api/v1/auth/register"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/players/": {
      "get": {
        "summary": "Listar atletas com paginação e busca",
        "tags": [
          "Atletas"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "nationality",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "GOALKEEPER",
                "DEFENDER",
                "MIDFIELDER",
                "FORWARD"
              ]
            },
            "in": "query",
            "name": "position",
            "required": false
          },
          {
            "schema": {
              "default": 20,
              "type": "number",
              "minimum": 1,
              "maximum": 100
            },
            "in": "query",
            "name": "limit",
            "required": false
          },
          {
            "schema": {
              "default": 1,
              "type": "number",
              "minimum": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/players/compare": {
      "get": {
        "summary": "Comparador Raio-X de Atletas (Head-to-Head)",
        "tags": [
          "Atletas"
        ],
        "description": "Compara dois jogadores lado a lado na temporada atual ou específica, calculando métricas ofensivas, defensivas e apontando as vantagens estatísticas de cada um.",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "p1",
            "required": true,
            "description": "ID do primeiro atleta"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "p2",
            "required": true,
            "description": "ID do segundo atleta"
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "seasonId",
            "required": false,
            "description": "ID da temporada (opcional)"
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/players/{id}": {
      "get": {
        "summary": "Obter detalhes do atleta",
        "tags": [
          "Atletas"
        ],
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/players/{id}/statistics": {
      "get": {
        "summary": "Estatísticas e scouts detalhados do atleta",
        "tags": [
          "Atletas"
        ],
        "description": "Retorna o histórico de gols, assistências, rating, finalizações, passes-chave e scouts oficiais.",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "query",
            "name": "seasonId",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/players/{id}/career": {
      "get": {
        "summary": "Histórico completo de carreira e scouting do atleta",
        "tags": [
          "Atletas"
        ],
        "description": "Retorna o currículo completo temporada a temporada do atleta, com totais acumulados de carreira (gols, assistências, minutos, clean sheets, defesas e cartões).",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/players/{id}/fantasy": {
      "get": {
        "summary": "Histórico e médias de pontuação Fantasy & Cartola FC do Atleta",
        "tags": [
          "Atletas"
        ],
        "description": "Retorna as pontuações do atleta rodada a rodada, média na temporada, maior pontuação (mito da rodada) e menor pontuação.",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
    "/api/v1/players/{id}/heatmap": {
      "get": {
        "summary": "Mapa de calor e zonas de atuação em campo do atleta",
        "tags": [
          "Atletas"
        ],
        "description": "Retorna as coordenadas 2D dos toques do atleta na partida, proporção de atuação por terço do campo e alas.",
        "parameters": [
          {
            "schema": {
              "default": 1,
              "type": "number"
            },
            "in": "query",
            "name": "matchId",
            "required": false
          },
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/export/players": {
      "get": {
        "summary": "Exportar scouts e estatísticas de atletas em formato CSV",
        "tags": [
          "Exportação de Dados"
        ],
        "description": "Gera arquivo .csv para download com scouts completos de jogadores por liga ou temporada.",
        "parameters": [
          {
            "schema": {
              "default": 1,
              "type": "number"
            },
            "in": "query",
            "name": "seasonId",
            "required": false
          },
          {
            "schema": {
              "default": 50,
              "type": "number",
              "minimum": 1,
              "maximum": 200
            },
            "in": "query",
            "name": "limit",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/api/v1/scouting/players/{playerId}": {
      "get": {
        "summary": "Ficha técnica e relatório completo de scouting do atleta",
        "tags": [
          "Scouting & Olheiro de Talentos"
        ],
        "description": "Retorna a teia de atributos técnicos e físicos (0 a 100), nota de potencial máximo, jogador de estilo similar no mundo e recomendação do olheiro.",
        "parameters": [
          {
            "schema": {
              "type": "number"
            },
            "in": "path",
            "name": "playerId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
...
}